# swagger/angular-client/model/tutorial.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Tutorial { 
    description?: string;
    id?: number;
    published?: boolean;
    title?: string;
}